% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofactor.pca.cor.R
\name{cofactor.pca.cor}
\alias{cofactor.pca.cor}
\title{Test for correlations between user-specified cofactors and principal components calculated from genotype data.}
\usage{
cofactor.pca.cor(U, G)
}
\arguments{
\item{U}{A numeric matrix containing user-specified cofactors. Dimensions are n rows (individuals) by c columns (cofactors).}

\item{G}{A numeric matrix containing genotype data. Dimensions are n rows (individuals) by m columns (genetic markers).}
}
\value{
A numeric matrix containing user-specified cofactors and all principal components not correlated with the
 user-specified cofactors. Dimensions are n rows (individuals) by c columns (cofactors).

If U is unspecified, cofactor.pca.cor will carry out principal components analysis identically to the native R function prcomp(),
 and cofactor.pca.cor will return principal components scores in a numeric matrix with n rows (individuals) by c columns (principal components).
}
\description{
Test for correlations between user-specified cofactors and principal components calculated from genotype data.
}
