% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASbyGLM.R
\name{GWASbyGLM}
\alias{GWASbyGLM}
\title{Genome-wide association study using a generalized linear model.}
\usage{
GWASbyGLM(y, G, C, PC)
}
\description{
Genome-wide association study using a generalized linear model.
}
