% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM2020-package.R
\docType{data}
\name{GLM2020-package}
\alias{GLM2020-package}
\title{A tool for implementing genome-wide association analysis using a general linear model.}
\format{
An object of class \code{character} of length 1.
}
\usage{
GLM2020-package
}
\description{
This package offers several features for efficiently carrying out genome-wide association studies (GWAS) using general linear models (GLM). Users may input phenotypic and genotypic data, as well as user-specified cofactors (e.g., location collected) and principal components calculated from genotype data.
In addition to implementing GWAS using a GLM, this package offers functionality for automatically removing genotypic principal components that exhibit correlation with user-specified cofactors.
}
\keyword{datasets}
